refactor: 패키지명 코틀린 코딩 컨벤션에 맞춰 수정
