+++ src/main/kotlin/dslstudy/Language.kt
@@ -1,4 +1,4 @@
-package dsl_study
+package dslstudy
 
 data class Language(
     val name: String,
@@ -14,4 +14,15 @@ class LanguagesBuilder(
     }
 
     fun build(): List<Language> = languages.toList()
+}
+
+class A {
+    val a = 10
+}
+
+class B {
+    val a = 20
+    val b = 30
+
+    fun A.aTest() = a + b + this@B.a
 }
